code review nits
