+++ politeiawww/cmd/politeiavoter/config.go
@@ -63,16 +63,16 @@ var runServiceCommand func(string) error
 //
 // See loadConfig for details on the configuration load process.
 type config struct {
+	ListCommands     bool `short:"l" long:"listcommands" description:"List available commands"`
+	ShowVersion      bool `short:"V" long:"version" description:"Display version information and exit"`
+	Version          string
 	HomeDir          string `short:"A" long:"appdata" description:"Path to application home directory"`
-	ShowVersion      bool   `short:"V" long:"version" description:"Display version information and exit"`
 	ConfigFile       string `short:"C" long:"configfile" description:"Path to configuration file"`
 	LogDir           string `long:"logdir" description:"Directory to log output."`
 	TestNet          bool   `long:"testnet" description:"Use the test network"`
 	PoliteiaWWW      string `long:"politeiawww" description:"Politeia WWW host"`
 	Profile          string `long:"profile" description:"Enable HTTP profiling on given port -- NOTE port must be between 1024 and 65536"`
 	DebugLevel       string `short:"d" long:"debuglevel" description:"Logging level for all subsystems {trace, debug, info, warn, error, critical} -- You may also specify <subsystem>=<level>,<subsystem2>=<level>,... to set the log level for individual subsystems -- Use show to list available subsystems"`
-	Version          string
-	ListCommands     bool   `short:"l" long:"listcommands" description:"List available commands"`
 	WalletHost       string `long:"wallethost" description:"Wallet host"`
 	WalletCert       string `long:"walletgrpccert" description:"Wallet GRPC certificate"`
 	WalletPassphrase string `long:"walletpassphrase" description:"Wallet decryption passphrase"`
